requirements.txt
requests
pycryptodome
pyyaml
schedule
